User Role Refactoring
Added more conditional logic to seperate the responsibilities of buyers and sellers. Now a seller will need to create a second buyer account to make offers on properties. This cleans up the seller interface. The call to action on each page was updated to reflect the role of the user. TODO: create separate landing pages for buyers and sellers.